+++ src/app/events.js
@@ -1,4 +1,5 @@
-import { projectShow } from './projects';
+import projectShow from './projectDom';
+import { todoForm } from './todos';
 
 const newProject = document.getElementById('new-project');
 const showProject = document.getElementById('show-projects');
@@ -6,13 +7,13 @@ const newTodo = document.getElementById('new-todo');
 
 const projectForm = document.getElementById('project-form');
 const projectList = document.getElementById('project-list');
-const todoForm = document.getElementById('todo-form');
+const formTodo = document.getElementById('todo-form');
 const todoList = document.getElementById('todo-list');
 
 const elements = [
   [newProject, projectForm],
   [showProject, projectList],
-  [newTodo, todoForm],
+  [newTodo, formTodo],
 ];
 
 const createEvent = () => {
@@ -32,4 +33,8 @@ export default createEvent;
 
 showProject.addEventListener(('click'), () => {
   projectShow();
+});
+
+newTodo.addEventListener(('click'), () => {
+  todoForm();
 });
