+++ src/index.js
@@ -1,4 +1,12 @@
-import start from './app/main';
+import createEvent from './app/events';
+import projectForm from './app/projects';
+import projectShow from './app/projectDom';
+import { todoForm } from './app/todos';
+import { createOptions } from './app/todoDom';
 import './main.scss';
 
-start();
+createEvent();
+projectForm();
+projectShow();
+todoForm();
+createOptions();
